refactor: split VMap.vue into smaller modules
- Extract map layer definitions to src/constants/mapLayers.ts
- Extract geometry utilities (clipAndEnvelope, normalizeBbox) to src/utils/geom.ts
- Reduce VMap.vue from 456 to 236 lines focused on map lifecycle and events

Closes #29
